Move discovery test code into their correct modules
