+++ src/test/pythonEnvironments/base/info/environmentInfoService.functional.test.ts
@@ -6,16 +6,16 @@
 import * as assert from 'assert';
 import * as sinon from 'sinon';
 import { ImportMock } from 'ts-mock-imports';
-import { ExecutionResult } from '../../../client/common/process/types';
-import { IDisposableRegistry } from '../../../client/common/types';
-import { Architecture } from '../../../client/common/utils/platform';
-import { InterpreterInformation } from '../../../client/pythonEnvironments/base/info/interpreter';
-import { parseVersion } from '../../../client/pythonEnvironments/base/info/pythonVersion';
-import * as ExternalDep from '../../../client/pythonEnvironments/common/externalDependencies';
+import { ExecutionResult } from '../../../../client/common/process/types';
+import { IDisposableRegistry } from '../../../../client/common/types';
+import { Architecture } from '../../../../client/common/utils/platform';
+import { InterpreterInformation } from '../../../../client/pythonEnvironments/base/info/interpreter';
+import { parseVersion } from '../../../../client/pythonEnvironments/base/info/pythonVersion';
+import * as ExternalDep from '../../../../client/pythonEnvironments/common/externalDependencies';
 import {
     EnvironmentInfoServiceQueuePriority,
     getEnvironmentInfoService,
-} from '../../../client/pythonEnvironments/base/info/environmentInfoService';
+} from '../../../../client/pythonEnvironments/base/info/environmentInfoService';
 
 suite('Environment Info Service', () => {
     let stubShellExec: sinon.SinonStub;
