+++ src/test/pythonEnvironments/base/locators/composite/resolverUtils.unit.test.ts
@@ -18,7 +18,7 @@ import { buildEnvInfo } from '../../../../../client/pythonEnvironments/base/info
 import { InterpreterInformation } from '../../../../../client/pythonEnvironments/base/info/interpreter';
 import { parseVersion } from '../../../../../client/pythonEnvironments/base/info/pythonVersion';
 import { TEST_LAYOUT_ROOT } from '../../../common/commonTestConstants';
-import { assertEnvEqual } from '../../../discovery/locators/envTestUtils';
+import { assertEnvEqual } from '../envTestUtils';
 import { Architecture } from '../../../../../client/common/utils/platform';
 import {
     AnacondaCompanyName,
