+++ src/test/pythonEnvironments/base/locators/lowLevel/condaLocator.unit.test.ts
@@ -4,11 +4,11 @@
 import * as fsapi from 'fs-extra';
 import * as path from 'path';
 import * as sinon from 'sinon';
-import { PythonReleaseLevel, PythonVersion } from '../../../../client/pythonEnvironments/base/info';
-import * as externalDeps from '../../../../client/pythonEnvironments/common/externalDependencies';
-import { getPythonVersionFromConda } from '../../../../client/pythonEnvironments/common/environmentManagers/conda';
-import { TEST_DATA_ROOT } from '../../common/commonTestConstants';
-import { assertVersionsEqual } from './envTestUtils';
+import { PythonReleaseLevel, PythonVersion } from '../../../../../client/pythonEnvironments/base/info';
+import * as externalDeps from '../../../../../client/pythonEnvironments/common/externalDependencies';
+import { getPythonVersionFromConda } from '../../../../../client/pythonEnvironments/common/environmentManagers/conda';
+import { TEST_DATA_ROOT } from '../../../common/commonTestConstants';
+import { assertVersionsEqual } from '../envTestUtils';
 
 suite('Conda Python Version Parser Tests', () => {
     let readFileStub: sinon.SinonStub;
