+++ src/test/pythonEnvironments/base/locators/lowLevel/globalVirtualEnvironmentLocator.testvirtualenvs.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 
 import * as path from 'path';
-import { GlobalVirtualEnvironmentLocator } from '../../../../client/pythonEnvironments/base/locators/lowLevel/globalVirtualEnvronmentLocator';
-import { TEST_LAYOUT_ROOT } from '../../common/commonTestConstants';
+import { GlobalVirtualEnvironmentLocator } from '../../../../../client/pythonEnvironments/base/locators/lowLevel/globalVirtualEnvronmentLocator';
+import { TEST_LAYOUT_ROOT } from '../../../common/commonTestConstants';
 import { testLocatorWatcher } from './watcherTestUtils';
 
 suite('GlobalVirtualEnvironment Locator', async () => {
