+++ src/test/pythonEnvironments/base/locators/lowLevel/poetryLocator.testvirtualenvs.ts
@@ -4,14 +4,14 @@
 import { expect } from 'chai';
 import * as path from 'path';
 import * as sinon from 'sinon';
-import { ExecutionResult, ShellOptions } from '../../../../client/common/process/types';
-import { PythonEnvKind } from '../../../../client/pythonEnvironments/base/info';
-import { BasicEnvInfo, ILocator } from '../../../../client/pythonEnvironments/base/locator';
-import { getEnvs } from '../../../../client/pythonEnvironments/base/locatorUtils';
-import * as externalDependencies from '../../../../client/pythonEnvironments/common/externalDependencies';
-import { PoetryLocator } from '../../../../client/pythonEnvironments/base/locators/lowLevel/poetryLocator';
-import { EXTENSION_ROOT_DIR_FOR_TESTS } from '../../../constants';
-import { TEST_LAYOUT_ROOT } from '../../common/commonTestConstants';
+import { ExecutionResult, ShellOptions } from '../../../../../client/common/process/types';
+import { PythonEnvKind } from '../../../../../client/pythonEnvironments/base/info';
+import { BasicEnvInfo, ILocator } from '../../../../../client/pythonEnvironments/base/locator';
+import { getEnvs } from '../../../../../client/pythonEnvironments/base/locatorUtils';
+import * as externalDependencies from '../../../../../client/pythonEnvironments/common/externalDependencies';
+import { PoetryLocator } from '../../../../../client/pythonEnvironments/base/locators/lowLevel/poetryLocator';
+import { EXTENSION_ROOT_DIR_FOR_TESTS } from '../../../../constants';
+import { TEST_LAYOUT_ROOT } from '../../../common/commonTestConstants';
 import { testLocatorWatcher } from './watcherTestUtils';
 
 suite('Poetry Watcher', async () => {
