+++ src/test/pythonEnvironments/base/locators/lowLevel/poetryLocator.unit.test.ts
@@ -3,15 +3,15 @@
 
 import * as path from 'path';
 import * as sinon from 'sinon';
-import { PythonEnvKind } from '../../../../client/pythonEnvironments/base/info';
-import * as externalDependencies from '../../../../client/pythonEnvironments/common/externalDependencies';
-import * as platformUtils from '../../../../client/common/utils/platform';
-import { getEnvs } from '../../../../client/pythonEnvironments/base/locatorUtils';
-import { PoetryLocator } from '../../../../client/pythonEnvironments/base/locators/lowLevel/poetryLocator';
-import { TEST_LAYOUT_ROOT } from '../../common/commonTestConstants';
-import { assertBasicEnvsEqual } from './envTestUtils';
-import { ExecutionResult, ShellOptions } from '../../../../client/common/process/types';
-import { createBasicEnv } from '../../base/common';
+import { PythonEnvKind } from '../../../../../client/pythonEnvironments/base/info';
+import * as externalDependencies from '../../../../../client/pythonEnvironments/common/externalDependencies';
+import * as platformUtils from '../../../../../client/common/utils/platform';
+import { getEnvs } from '../../../../../client/pythonEnvironments/base/locatorUtils';
+import { PoetryLocator } from '../../../../../client/pythonEnvironments/base/locators/lowLevel/poetryLocator';
+import { TEST_LAYOUT_ROOT } from '../../../common/commonTestConstants';
+import { assertBasicEnvsEqual } from '../envTestUtils';
+import { ExecutionResult, ShellOptions } from '../../../../../client/common/process/types';
+import { createBasicEnv } from '../../common';
 
 suite('Poetry Locator', () => {
     let shellExecute: sinon.SinonStub;
