+++ src/test/pythonEnvironments/base/locators/lowLevel/workspaceVirtualEnvLocator.testvirtualenvs.ts
@@ -5,7 +5,7 @@ import * as path from 'path';
 import { PythonEnvKind } from '../../../../../client/pythonEnvironments/base/info';
 import { WorkspaceVirtualEnvironmentLocator } from '../../../../../client/pythonEnvironments/base/locators/lowLevel/workspaceVirtualEnvLocator';
 import { TEST_LAYOUT_ROOT } from '../../../common/commonTestConstants';
-import { testLocatorWatcher } from '../../../discovery/locators/watcherTestUtils';
+import { testLocatorWatcher } from './watcherTestUtils';
 
 suite('WorkspaceVirtualEnvironment Locator', async () => {
     const testWorkspaceFolder = path.join(TEST_LAYOUT_ROOT, 'workspace', 'folder1');
