+++ src/test/pythonEnvironments/common/environmentManagers/simplevirtualenvs.unit.test.ts
@@ -15,8 +15,8 @@ import {
     isVirtualenvEnvironment,
     isVirtualenvwrapperEnvironment,
 } from '../../../../client/pythonEnvironments/common/environmentManagers/simplevirtualenvs';
-import { TEST_DATA_ROOT, TEST_LAYOUT_ROOT } from '../../common/commonTestConstants';
-import { assertVersionsEqual } from './envTestUtils';
+import { TEST_DATA_ROOT, TEST_LAYOUT_ROOT } from '../commonTestConstants';
+import { assertVersionsEqual } from '../../base/locators/envTestUtils';
 
 suite('isVenvEnvironment Tests', () => {
     const pyvenvCfg = 'pyvenv.cfg';
